third_party/blink/web_tests/fast/events/custom-event-expected.txt

PASS: arguments[0].detail.result of the CustomEvent should be 'success' and is.
PASS: arguments[0].detail of the CustomEvent should be 'success' and is.
